\name{randomAlphabet-package}
\alias{randomAlphabet-package}
\alias{randomAlphabet}
\docType{package}
\title{
\packageTitle{randomAlphabet}
}
\description{
\packageDescription{randomAlphabet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{randomAlphabet}
\packageIndices{randomAlphabet}

}
\author{
\packageAuthor{randomAlphabet}

Maintainer: \packageMaintainer{randomAlphabet}
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{

}
